move `resolve_fingerprint_variable` to fingerprinting utils module
