+++ src/sentry/grouping/fingerprinting/utils.py
@@ -5,8 +5,10 @@ import re
 from collections.abc import Mapping
 from typing import Any, Literal, NotRequired, TypedDict
 
+from sentry.db.models.fields.node import NodeData
 from sentry.stacktraces.functions import get_function_name_for_frame
 from sentry.stacktraces.platform import get_behavior_family_for_platform
+from sentry.stacktraces.processing import get_crash_frame_from_event_data
 from sentry.utils.event_frames import find_stack_frames
 from sentry.utils.safe import get_path
 from sentry.utils.tag_normalization import normalized_sdk_tag_from_event
@@ -197,3 +199,79 @@ def get_fingerprint_type(
             else "custom"
         )
     )
+
+
+def resolve_fingerprint_variable(
+    variable_key: str,
+    event_data: NodeData | Mapping[str, Any],
+    use_legacy_unknown_variable_handling: bool,
+) -> str | None:
+    if variable_key == "transaction":
+        return event_data.get("transaction") or "<no-transaction>"
+
+    elif variable_key == "message":
+        message = (
+            get_path(event_data, "logentry", "formatted")
+            or get_path(event_data, "logentry", "message")
+            or get_path(event_data, "exception", "values", -1, "value")
+        )
+        return message or "<no-message>"
+
+    elif variable_key in ("type", "error.type"):
+        exception_type = get_path(event_data, "exception", "values", -1, "type")
+        return exception_type or "<no-type>"
+
+    elif variable_key in ("value", "error.value"):
+        value = get_path(event_data, "exception", "values", -1, "value")
+        return value or "<no-value>"
+
+    elif variable_key in ("function", "stack.function"):
+        frame = get_crash_frame_from_event_data(event_data)
+        func = frame.get("function") if frame else None
+        return func or "<no-function>"
+
+    elif variable_key in ("path", "stack.abs_path"):
+        frame = get_crash_frame_from_event_data(event_data)
+        abs_path = frame.get("abs_path") or frame.get("filename") if frame else None
+        return abs_path or "<no-abs-path>"
+
+    elif variable_key == "stack.filename":
+        frame = get_crash_frame_from_event_data(event_data)
+        filename = frame.get("filename") or frame.get("abs_path") if frame else None
+        return filename or "<no-filename>"
+
+    elif variable_key in ("module", "stack.module"):
+        frame = get_crash_frame_from_event_data(event_data)
+        module = frame.get("module") if frame else None
+        return module or "<no-module>"
+
+    elif variable_key in ("package", "stack.package"):
+        frame = get_crash_frame_from_event_data(event_data)
+        pkg = frame.get("package") if frame else None
+        if pkg:
+            # If the package is formatted as either a POSIX or Windows path, grab the last segment
+            pkg = pkg.rsplit("/", 1)[-1].rsplit("\\", 1)[-1]
+        return pkg or "<no-package>"
+
+    elif variable_key == "level":
+        return event_data.get("level") or "<no-level>"
+
+    elif variable_key == "logger":
+        return event_data.get("logger") or "<no-logger>"
+
+    elif variable_key.startswith("tags."):
+        # Turn "tags.some_tag" into just "some_tag"
+        requested_tag = variable_key[5:]
+        for tag_name, tag_value in event_data.get("tags") or ():
+            if tag_name == requested_tag and tag_value is not None:
+                return tag_value
+        return "<no-value-for-tag-%s>" % requested_tag
+    else:
+        # TODO: Once we have fully transitioned off of the `newstyle:2023-01-11` grouping config, we
+        # can remove `use_legacy_unknown_variable_handling` and just return the string. (At that
+        # point we can also change the return type of this function to just be `str`.)
+        return (
+            None
+            if use_legacy_unknown_variable_handling
+            else "<unrecognized-variable-%s>" % variable_key
+        )
